subo cosas que iba a subir en la noche, sumo al calculo dispositivo.horasDeUso() ,estaba esperando sorry

+++ src/test/java/sensorTest/SensorTest.java
@@ -64,7 +64,7 @@ public class SensorTest {
 		reglaSimple.agregarActuador(actuadorLuz);
 		reglaSimple.agregarActuador(actuadorTemperatura);
 		
-		sensor.obtenerMedicion(new Medicion(10));		
+		sensor.obtenerMedicion(new Medicion((double)10));
 	}
 	
 }
